Implement Cloud Foundry service broker
Using a Sinatra app to implement the service broker interface.

Can be started using `bundle exec rackup`

With the service broker model we only look for apps which are explicitly bound
to the Prometheus service. This means if an app has been stored we need to
retain its configuration.

These changes make it so when we notice an app has been stopped we move its
configuration into a separate directory from the running apps. When apps are
unbound we remove all configuration.