Add a new filestore_manager to support S3
Config files can be stored locally (for tests) or in S3. Previously we
were accessing the config files directly whereas this updates to use a
common interface for moving / adding /deleting config shared between the
s3_manager.rb and the local_manager.rb.

The manager used is selected by a factory which inspects the environment
variable ENVIRONMENT to choose which config file manager to use.

Currently local_manager is only for tests.