Pin ruby version
We don't pin our ruby version anywhere.  This means we're at the mercy
of whatever the buildpack chooses to offer us.  We should probably be
explicit about which ruby version we use.

This adds a `.ruby-version` file (used by rbenv, for local
development) and a `ruby` clause to the Gemfile (used by bundler and
by ruby_buildpack).  I use the twiddle-wakka (~>) operator to allow
later 2.4.x versions as they become available in ruby_buildpack, as
suggested in the ruby_buildpack docs:
https://docs.cloudfoundry.org/buildpacks/ruby/index.html

There is a risk that `.ruby-version` and `Gemfile` get out of sync,
but if that happens, it will fail loudly when bundler complains that
the running ruby version isn't the same as what Gemfile requires, so I
don't think it's a big problem.

My understanding is that currently we get given 2.4.3 by the
buildpack.  I worked this out by running `cf buildpacks` to see what
version of ruby_buildpack is available and viewing the ruby_buildpack
version history:
https://github.com/cloudfoundry/ruby-buildpack/releases

As a result, I chose to pin to the version we're already running.