change instance label to include app guid
For zero-downtime deployments, the new and old apps will run
concurrently, but with different guids.  We will also scrape them
concurrently.  But we assign the same `instance` label to them.

We should therefore produce an `instance` label that is unique per
guid:instance-index pair.

This may break promql expressions which make assumptions about the old
shape of the instance label.