strip "-venerable" suffix from job label
The autopilot cloud foundry deploy tool works by renaming the existing
instance of an app "foo" to "foo-venerable", before deploying the new
instance as an app called "foo".

We actually want all of these apps to have the same job name, so that
we can write PromQL queries that show all of the concurrent instances
of an app (old and new), and also so that our app metrics don't get
renamed when autopilot renames the app.

It's a bit of a hack, but we can just hardcode the service broker to
strip "-venerable" suffixes from app names so they all have the same
job.