Fix link to RE docs
We provide a link in the service description for the GDS Prometheus
service broker. This link had broken recently as we changed our
section headings and the anchor no longer pointed to a existant
heading.

I could have fixed it to point at the new anchor for our section
but thought this may continue to change in the future so went
for the easiest long run approach on the assumption that once
users make it to the RE docs they will be able to find the content
about our service.