Don't add space label to paas-prometheus-exporter targets
In the same way we don't add the space label to paas-metric-exporter
targets, we don't want to add the space label to
paas-prometheus-exporter targets.

In time we will likely transition away from our users using the
paas-metric-exporter as a prometheus endpoint and therefore can
refactor out support for this and do some more appropriate
renaming. We've kept our changes as simple as possible for
the moment.