Correct uses of `detected_start_command: nil`
By curling the cf api (cf curl v2/apps), you see that every
app has a `detected_start_command` and if an app doesn't then
it is still left as an empty string rather than nil. The content
of the API response is then parsed using JSON.parse which does
not change it to nil either.