Remove confusing implication in test
These targets had their detected_start_command as
`./bin/paas-metric-exporter`. Because this start command is
a special case it implies that this test might be doing something
to do with that special behaviour but in fact this is not the case
and it appears to have just been used as an example.

I have changed it to something more generic to remove the
implication that this test is doing something special based
on the detected start command.