Remove integration assertion
We are already testing the behaviour of `target.paas_metric_exporter?`
in unit tests in target_spec.rb. It is not beneficial to
test it again in `parser_spec.rb`. We instead just want to know
that the `detected_start_command` field is being parsed correctly.

We have integration tests elsewhere for the dropping of space labels
for paas metric exporter targets.