Implement HTTP GET/HEAD request caching on API calls
This service was generating a lot of HTTP requests every 60 seconds, causing
noticable load on the API. In the future the API will be rate-limited to
prevent accidental denials of service, and this could severely impact this
service. To mitigate that, we implement caching of HTTP GET and HEAD requests,
so that the service will be less likely to hit the request rate limits.