Fix bug with apps with only internal routes
This fixes a bug where an app which only had internal routes would not
be handled correctly.  Admittedly it's not clear what we *can* do if
there aren't any public routes but we should at least not cause
"undefined method dig for NilClass" exceptions.

We should also log when we're guessing a route because we're almost
certainly wrong.