Disable Style/StderrPuts cop
- This was complaining about:

```
lib/cf_app_discovery/service_broker.rb:64:9: C: Style/StderrPuts: Use warn instead of STDERR.puts to allow such output to be disabled. (https://rubystyle.guide#warn)
        STDERR.puts "Entity key unavailable, please check data structure:\n #{service_binding}"
```

and rather than potentially break an app that I've not worked on for
a while, I'll leave it for someone else to re-enable in the future.