Replace custom VCAP_SERVICES code with env-map buildpack
We have some custom code for slurping credentials out of a
user-provided service in VCAP_SERVICES and dumping them into separate
environment variables.  This is a bit flaky in a few ways:

 - we don't explicitly say which credentials we're expecting in the
   user-provided service; we just dump everything we find into the
   environment
 - it's custom code we have to understand in order to maintain this
   repo

This commit replaces them with @andy-paine's env-map-buildpack, which
allows you to pull arbitrary `yq` expressions out of VCAP_SERVICES
into environment variables using an `env-map.yml` file.

Further, in future if we decide to use the [S3 broker][], it will be
straightforward to use `env-map.yml` to extract credentials from the
S3 broker service instead of from the user-provided service.  With the
current code, it would not be easy to perform this migration.

[S3 broker]: https://docs.cloud.service.gov.uk/deploying_services/s3/